dev-util/gdmd: call gcc-config in pkg_post*, update EAPI 7 -> 8
For the gdmd executable to be available in $PATH gcc-config needs to be
called, only if the currently selected config is the same one as the
slot of the package. Similarly, the removal of the package should
trigger a call to gcc-config to remove the dangling symlink.

Logic has been taken from toolchain.eclass and adapted to treat only the
basic cases as this package doesn't have a high importance to the
system.

Signed-off-by: Horodniceanu Andrei <a.horodniceanu@proton.me>