*/*: change dlang-filter-dflags
Change the dlang-filter-dflags function to:
1. Take a compiler name instead of an implementation pattern as the
   first argument.
2. Not modify DMDW_DCFLAGS if it is set.
3. Fix the check that DCFLAGS is set when deciding if it should
   be updated. The previous check always succeeded.

The rationale for these changes is, as follows:

The previous implementation picked three random version of D
implementations, one for each compiler, and checked those against the
supplied <pattern> argument. This, at a first glance, already sounds
weird so the code quality is iffy.

If that check succeeded then the variable associated with the
implementation would be stripped accordingly. Since the picked
implementations are unknown to the consumers of the eclass it means that
the supplied patterns must include a glob, like `dmd*'. If this is the
case it is simpler for the consumer to just supply `dmd' and have the
same stripping be carried on without any other checks.

This would remove the ability to specify `*' as a pattern meaning to
filter the flags from every variable. Since the flags are usually
different across compilers, especially dmd and ldc2 vs gdc, it is an
acceptable removal, even though `*' is used by dev-util/dub. The ebuild
needs to filter some flags that are only problematic with ldc2 so the
pattern can be changed to `ldc' without consequences.

The second point is related to the ways the eclass provides to access
user-provided flags for D compilers. The first such way is the DCFLAGS
variable. It is expected, then, that when filter-dflags finishes the
DCFLAGS variable will contain the filtered flags. This same logic
doesn't apply to DMDW_DCFLAGS, however, since ebuilds are expected to
use the dlang_get_dmdw_dcflags function. It makes no sense to modify the
internal variable used by _dlang_export. Worse, if a user saved
dlang_get_dmdw_dcflags in a variable called DMDW_DCFLAGS, like what
dmd-r1.eclass does, then filter-dflags would modify that variable which
can be unexpected if one doesn't know the internals of the eclass. For
this reason this behavior has been removed.

Additionally, I want to provide another solution with its pros and cons
and the reasons why it is not satisfactory.

Instead of the check on the three random implementation the check can be
carried against the EDC variable. This way we retain the old behavior of
accepting globs for the <pattern> but we also allow specifying specific
versions. This has practical applicability with the gdc-13 bug
preventing avx optimizations. Ebuilds could do
`dlang-filter-dflags gdc-13 -march=native` and have that flag removed
for the problematic implementation. The only problem with this comes
when trying to use filter-dflags from a multi-implementation ebuild.

Because of the requirement that EDC is set (so that <pattern> can be
checked against it) dlang-filter-dflags would be needed to be called
inside dlang_foreach_impl. This is
- inefficient
- unintuitive
  Because someone might think that filter-dflags only modifies the
  temporary variable DCFLAGS. One would also lose the ability to specify
  specific implementation as the affected flags variable is the same.

It is not possible to change filter-dflags to only modify the DCFLAGS
variable as dlang_get_dmdw_dcflags should return the filtered flags and
that function always regenerates the value from the user flags.

In order to maintain feature-parity between filter-dflags across
dlang-r1.eclass and dlang-single.eclass the first solution is preferred.
An ebuild is expected to use dlang-filter-flags function as follows:

```
 # dlang-single
 dlang-filter-dflags gdc --your-flag
 # filter --your-flag from GDCFLAGS.
 # $DCFLAGS and $(dlang_get_dmww_dcflags) no longer contain --your-flag
```

```
 # dlang-single
 [[ ${EDC} == gdc-13 ]] && dlang-filter-dflags gdc --your-flag
 # Only filter the flag on a specific implementation
 # Required the user to duplicate `gdc'
```

```
 # dlang-multi
 src_configure() {
   dlang-filter-dflags gdc --your-flag
   dlang-filter-dflags ldc --other

   dlang_foreach_impl [...]
   # Inside dlang_foreach_impl DCFLAGS will not contain --your-flag
   # for gdc implementations or --other for ldc implementations.
   # Same thing applies to $(dlang_get_dmdw_dcflags)
 }
```

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>